requirements.txt
# PDF Processing
PyMuPDF>=1.24.0
Pillow>=10.0.0

# AI/LLM APIs
google-generativeai>=0.8.0
openai>=1.46.0
openrouter>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-docx>=1.1.2
pydantic>=2.8.0

# Development/Notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
